// src/app/core/enums/statistics.enum.ts
export enum StatisticsEnum {
  //Kilometer Ranking
  KilometerStatistics = 1,
  //logbook
  logbookStatistics = 2,
  //logbook of a single boat
  logbookSingleBoat = 3,
  //Commerzbank Ranking
  CommerzbankStatistics = 4,
  //Venezianer Ranking
  VenezianerStatistics = 5,
  //Venezianer Ranking
  VenezianerStatisticsPersonal = 6,
}